Remove `dolfinx_utils` (#1982)
* Remove a skip

* Simplify skips

* Remove dolfinx_utils

* Add missing import

* Remove unused functions

* Avoid accessing PETSc vector in a test

* Remove pointless test

* Fix parallel skip
