+++ python/test/unit/io/test_xdmf_function.py
@@ -15,12 +15,9 @@ from dolfinx.fem import (Function, FunctionSpace, TensorFunctionSpace,
 from dolfinx.io import XDMFFile
 from dolfinx.mesh import (CellType, create_unit_cube, create_unit_interval,
                           create_unit_square)
-from dolfinx_utils.test.fixtures import tempdir
 
 from mpi4py import MPI
 from petsc4py import PETSc
-
-assert (tempdir)
 
 # Supported XDMF file encoding
 if MPI.COMM_WORLD.size > 1:
@@ -40,15 +37,6 @@ def mesh_factory(tdim, n):
         return create_unit_square(MPI.COMM_WORLD, n, n)
     elif tdim == 3:
         return create_unit_cube(MPI.COMM_WORLD, n, n, n)
-
-
-@pytest.fixture
-def worker_id(request):
-    """Return worker ID when using pytest-xdist to run tests in parallel"""
-    if hasattr(request.config, 'slaveinput'):
-        return request.config.slaveinput['slaveid']
-    else:
-        return 'master'
 
 
 # --- Function
