+++ python/test/unit/mesh/test_face.py
@@ -11,7 +11,6 @@ from dolfinx import cpp as _cpp
 from dolfinx.cpp.mesh import cell_normals
 from dolfinx.mesh import (create_unit_cube, create_unit_square,
                           locate_entities_boundary)
-from dolfinx_utils.test.skips import skip_in_parallel
 
 from mpi4py import MPI
 
@@ -27,7 +26,7 @@ def square():
 
 
 @pytest.mark.skip("volume_entities needs fixing")
-@skip_in_parallel
+@pytest.mark.skip_in_parallel
 def test_area(cube, square):
     """Iterate over faces and sum area."""
