+++ python/test/unit/mesh/test_periodic_boundary_computation.py
@@ -8,7 +8,6 @@
 
 # from dolfinx import create_unit_square
 # from dolfinx.cpp.mesh import PeriodicBoundaryComputation
-# from dolfinx_utils.test.skips import skip_in_parallel
 
 
 # def periodic_boundary(x):
@@ -20,7 +19,7 @@
 #     return create_unit_square(MPI.COMM_WORLD, 4, 4)
 
 
-# @skip_in_parallel
+# @pytest.mark.skip_in_parallel
 # def test_ComputePeriodicPairs(mesh):
 #     # Verify that correct number of periodic pairs are computed
 #     vertices = PeriodicBoundaryComputation.compute_periodic_pairs(
@@ -31,7 +30,7 @@
 #     assert len(edges) == 4
 
 
-# @skip_in_parallel
+# @pytest.mark.skip_in_parallel
 # def test_MastersSlaves(mesh):
 #     # Verify that correct number of masters and slaves are marked
 #     mf = PeriodicBoundaryComputation.masters_slaves(mesh, periodic_boundary, 0, np.finfo(float).eps)
